EC-579 - separating out the editing and questions functions
